requirements.txt
django-ckeditor-5==0.2.12
pillow==10.3.0
psycopg2==2.9.9
python-slugify==8.0.4
django-select2==8.1.2
